refactor(85-02): extract EndgameScoreOverTimeChart to its own file
- Move chart + useIsMobile hook + SCORE_BAND_CLASS + timeline-only constants
  out of EndgamePerformanceSection.tsx (D-09)
- Update import in Endgames.tsx and the chart test file
- Section function preserved for Plan 04 to delete cleanly
