+++ frontend/src/pages/Endgames.tsx
@@ -18,7 +18,8 @@ import { InfoPopover } from '@/components/ui/info-popover';
 import { FilterPanel, DEFAULT_FILTERS, areFiltersEqual, FILTER_DOT_FIELDS } from '@/components/filters/FilterPanel';
 import { useFilterStore } from '@/hooks/useFilterStore';
 import { EndgameWDLChart } from '@/components/charts/EndgameWDLChart';
-import { EndgamePerformanceSection, EndgameScoreOverTimeChart } from '@/components/charts/EndgamePerformanceSection';
+import { EndgamePerformanceSection } from '@/components/charts/EndgamePerformanceSection';
+import { EndgameScoreOverTimeChart } from '@/components/charts/EndgameScoreOverTimeChart';
 import { EndgameStartVsEndSection } from '@/components/charts/EndgameStartVsEndSection';
 import { EndgameConvRecovChart } from '@/components/charts/EndgameConvRecovChart';
 import { EndgameScoreGapSection } from '@/components/charts/EndgameScoreGapSection';
